pyproject.toml
[project]
name = "function grapher"
version = "0.1.0"
description = "An interactive Python application for plotting a function, displaying a formula in LaTeX, and calculating the area under a curve using tkinter, matplotlib, and sympy."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "fonttools==4.57.0",
    "kiwisolver==1.4.8",
    "matplotlib==3.10.1",
    "mpmath==1.3.0",
    "numpy==2.2.4",
    "packaging==24.2",
    "pillow==11.1.0",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "six==1.17.0",
    "sympy==1.13.3",
]
